% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_amino_acids_dataframe.R
\name{all_amino_acids_dataframe}
\alias{all_amino_acids_dataframe}
\title{Create a dataframe with counts and frequency of all canonical amino acids}
\usage{
all_amino_acids_dataframe(file, sort_by = "seq_name", decreasing = TRUE)
}
\arguments{
\item{file}{fasta file containing amino acid sequences}

\item{sort_by}{the column in the final dataframe used for sorting}

\item{decreasing}{how to sort the final dataframe (TRUE is the default)}
}
\description{
Create a dataframe with counts and frequency of all canonical amino acids
}
