% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_acid_definitions.R
\docType{data}
\name{amino_acid_definitions}
\alias{amino_acid_definitions}
\title{Definition of amino acids by one letter.}
\format{
An object of class \code{list} of length 20.
}
\usage{
amino_acid_definitions
}
\description{
A list of amino acids.
}
\keyword{datasets}
