% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_acid_groups.R
\docType{data}
\name{amino_acid_groups}
\alias{amino_acid_groups}
\title{Definition of amino acid groups}
\format{
An object of class \code{list} of length 8.
}
\usage{
amino_acid_groups
}
\description{
A list of amino acid groups.
}
\keyword{datasets}
