% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_max_aa_per_100.R
\name{calculate_max_aa_per_100}
\alias{calculate_max_aa_per_100}
\title{Amino acid density score to identify a span of 100 amino acids with the highest count of an amino acid of interest}
\usage{
calculate_max_aa_per_100(
  file,
  amino_acid,
  sort_by = "max_aa_count",
  decreasing = TRUE
)
}
\arguments{
\item{file}{dataframe of amino acid sequences}

\item{sort_by}{the column in the final dataframe used for sorting}

\item{decreasing}{how to sort the final dataframe (TRUE is the default)}

\item{target_aa}{amino acid of interest}
}
\description{
Amino acid density score to identify a span of 100 amino acids with the highest count of an amino acid of interest
}
