% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aa_heatmap.R
\name{create_aa_heatmap}
\alias{create_aa_heatmap}
\title{Create an amino acid density heatmap}
\usage{
create_aa_heatmap(file)
}
\arguments{
\item{file}{fasta file containing amino acid sequences}
}
\description{
Create an amino acid density heatmap
}
