% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_amino_acid_frequency.R
\name{plot_amino_acid_frequency}
\alias{plot_amino_acid_frequency}
\title{Frequency distribution of a single peptide within a sequence}
\usage{
plot_amino_acid_frequency(sequence)
}
\arguments{
\item{sequence}{amino acid sequence as a string to be analyzed}
}
\description{
Frequency distribution of a single peptide within a sequence
}
