% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_faa.R
\name{read_faa}
\alias{read_faa}
\title{The following is from the read_faa function from the ampir package
citation: Legana C H W Fingerhut, David J Miller, Jan M Strugnell, Norelle L Daly, Ira R Cooke, ampir: an R package for fast genome-wide prediction of antimicrobial peptides, Bioinformatics, Volume 36, Issue 21, November 2020, Pages 5262–5263, https://doi.org/10.1093/bioinformatics/btaa653
Read FASTA amino acids file into a dataframe
This function reads a FASTA amino acids file into a dataframe
read_faa}
\usage{
read_faa(file = NULL)
}
\arguments{
\item{file}{file path to the FASTA format file containing the protein sequences
Additional details... returns a dataframe containing the sequence name (seq_name) and sequence (seq_aa) columns}
}
\description{
This function was adapted from \code{read.fasta.R} by Jinlong Zhang (jinlongzhang01@gmail.com) for the phylotools package (http://github.com/helixcn/phylotools)
file path to the FASTA format file containing the protein sequences
Dataframe containing the sequence name (seq_name) and sequence (seq_aa) columns
}
\note{
This function was adapted from \code{read.fasta.R} by Jinlong Zhang (jinlongzhang01@gmail.com) for the phylotools package (http://github.com/helixcn/phylotools)
}
