% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resido_amino_acid_group_count.R
\name{resido_amino_acid_group_count}
\alias{resido_amino_acid_group_count}
\title{function for counting a user-specified amino acid group
counts a user-specified amino acid group in the supplied amino acid sequence
output is a dataframe with the amino acid group count for each sequence}
\usage{
resido_amino_acid_group_count(
  file,
  amino_acid_group,
  sort_by = "seq_name",
  decreasing = FALSE,
  output_format = "dataframe"
)
}
\arguments{
\item{file}{a fasta file the contains the sequence name and amino acid sequence}

\item{amino_acid_group}{the amino acid group of interest}

\item{sort_by}{how the resulting output should be sorted sort_by and decreasing}

\item{decreasing}{if false will sort in increasing order if true will sort decreasing}

\item{output_format}{either "dataframe" or "csv"}
}
\description{
function for counting a user-specified amino acid group
counts a user-specified amino acid group in the supplied amino acid sequence
output is a dataframe with the amino acid group count for each sequence
}
