% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resido_amino_acid_percent.R
\name{resido_amino_acid_percent}
\alias{resido_amino_acid_percent}
\title{function for calculating percent of a user specified amino acid
counts and sorts by percent of a user-specified amino acid in the sequence
output is a dataframe with the percent amino acid for each sequence}
\usage{
resido_amino_acid_percent(
  file,
  amino_acid,
  sort_by = "seq_name",
  decreasing = FALSE,
  output_format = "dataframe"
)
}
\arguments{
\item{file}{a fasta file the contains the sequence name and amino acid sequence}

\item{amino_acid}{the amino acid of interest as one-letter code}

\item{sort_by}{how the resulting output should be sorted sort_by and decreasing}

\item{decreasing}{if false will sort in increasing order if true will sort decreasing}

\item{output_format}{either "dataframe" or "csv"}
}
\description{
function for calculating percent of a user specified amino acid
counts and sorts by percent of a user-specified amino acid in the sequence
output is a dataframe with the percent amino acid for each sequence
}
