% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_dataframe_as_csv.R
\name{save_dataframe_as_csv}
\alias{save_dataframe_as_csv}
\title{Function to save a dataframe as a CSV file}
\usage{
save_dataframe_as_csv(data, file = "output.csv", ...)
}
\arguments{
\item{data}{The dataframe to be saved}

\item{file}{The output file path}

\item{...}{Additional arguments passed to write.csv}
}
\description{
Function to save a dataframe as a CSV file
}
