% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_dataframe.R
\name{sort_dataframe}
\alias{sort_dataframe}
\title{a function to be used in conjunction with the count and percent functions that sorts the dataframe outputs}
\usage{
sort_dataframe(fasta_dataframe, sort_by = "seq_name", decreasing = FALSE)
}
\arguments{
\item{fasta_dataframe}{is the dataframe that is a result of the amino acid count/percentage functions}

\item{sort_by}{the column to be sorted in the dataframe and how to sort it}

\item{decreasing}{if false it will be sorted in increasing order if true it will be decreasing}
}
\description{
a function to be used in conjunction with the count and percent functions that sorts the dataframe outputs
}
